requirements.txt
fastapi
uvicorn[standard]
crewai
crewai-tools
python-dotenv
weasyprint